# docs/api_definition.yml
openapi: 3.0.3
info:
  title: Intelligent Personalized Care API
  version: 2.0.0
  description: Current backend HTTP contract. User roles and routes use PATIENT and PHYSIOTHERAPIST terminology.
    Catalogue exerciseID is a UUID; execution exerciseID is an assigned daily-exercise integer. Examples are synthetic.
    Protected operations use Bearer access tokens; physiotherapist-only operations generally also require approved
    credentials.
servers:
- url: http://localhost:8080
  description: Local development
- url: https://organic-byway-391719.ew.r.appspot.com
  description: Configured hosted endpoint; availability not verified
...
paths:
...
  /users/patients/{patientID}/plans/{planID}/daily_lists/{dailyListID}/exercises/{exerciseID}:
    get:
      tags:
      - Execution
      summary: Get patient video of exercise
      operationId: get_users_patients_patientID_plans_planID_daily_lists_dailyListID_exercises_exerciseID
      description: 'Returns video/mp4 for a camera result; a sensor-only result does not have a video. Authorization:
        authenticated user; ownership/relationship checks apply where relevant.'
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Success.
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
      parameters:
      - name: patientID
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Resource identifier.
      - name: planID
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: Assignment snapshot plan ID.
      - name: dailyListID
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: Resource identifier.
      - name: exerciseID
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: Assigned daily-exercise ID, not the catalogue UUID.
      - name: set
        in: query
        required: true
        schema:
          type: integer
          format: int32
    post:
      tags:
      - Execution
      summary: Post video of exercise
      operationId: post_users_patients_patientID_plans_planID_daily_lists_dailyListID_exercises_exerciseID
      description: 'Camera video submission for a prescribed set. Required parts: video and set. Optional metadata
        parts match the names below, including case-sensitive feedBack. Returns 204 with no body. Authorization:
        patient owner (or assigned physiotherapist on shared result/profile routes).'
      security:
      - bearerAuth: []
      responses:
        '204':
          description: No content.
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
      parameters:
      - name: patientID
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Resource identifier.
      - name: planID
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: Assignment snapshot plan ID.
      - name: dailyListID
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: Resource identifier.
      - name: exerciseID
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: Assigned daily-exercise ID, not the catalogue UUID.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - video
              - set
              properties:
                video:
                  type: string
                  format: binary
                set:
                  type: integer
                  format: int32
                feedBack:
                  type: string
                executionMode:
                  type: string
                withLoad:
                  type: boolean
                loadValue:
                  type: number
                  format: float
                  minimum: 0
                  exclusiveMinimum: true
                  description: Positive external weight in kg, required when withLoad is true; omitted otherwise.
                loadUnit:
                  type: string
                  enum: [kg]
                  description: Required as kg when withLoad is true; omitted otherwise.
                executionScore:
                  type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Problem:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
          format: int32
      required:
      - title
      - status
      example:
        title: Forbidden request
        status: 400
...
  responses:
    Error400:
      description: Invalid request, profile or prescription
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Invalid request, profile or prescription
            status: 400
    Error401:
      description: Missing, invalid or expired credentials
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Missing, invalid or expired credentials
            status: 401
    Error403:
      description: Wrong role, owner or patient relationship
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Wrong role, owner or patient relationship
            status: 403
    Error404:
      description: Resource or result not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Resource or result not found
            status: 404
    Error409:
      description: Conflicting record or occupied exercise set
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Conflicting record or occupied exercise set
            status: 409